requirements.txt
# Microsoft Agent Framework (pre-release)
# IMPORTANT: Install with --pre flag: pip install -r requirements.txt --pre
agent-framework==1.0.0b260212
agent-framework-azure-ai==1.0.0b260212

# Environment variable management
python-dotenv>=1.0.0

# Async support for Jupyter notebooks
nest-asyncio>=1.5.0

# Azure authentication
azure-identity>=1.15.0

# Data validation and settings management (required by agent-framework)
pydantic>=2.0.0

# For visualizing agent interactions (optional, but recommended for development)
graphviz>=0.20.0

# Optional: For enhanced development experience
# jupyter>=1.0.0
# ipykernel>=6.29.0
